feat(fab_game): graded-failure mechanism for purification — the edge-loaded Na ring
The Na→Q_ox→V_t consequence was an all-or-nothing cliff: a uniform Na shift
moves every die's V_t together, so the whole wafer flips between Na≈3e15 and
8e15 cm⁻³, and the decade-spaced feedstock grades (plus ~1/k-per-pass refining)
jump right over that half-decade band. Grade choice was binary — EGS=100%,
solar=0% — with no graded "some dies fail" middle the gradual-failure policy
wants.

Model mobile-ion Na as edge-loaded (the real handling/edge-bead/furnace radial
non-uniformity): Variation.na_factor(r) = 1 + na_edge_boost·r² (boost=2.5, a
flagged ~3.5× rim), applied per-die at the device step via
pipeline._die_contamination. A marginal feed now kills an outboard RING (~46%
yield, kills on the rim) where a uniform model called a clean pass — the honest
ok→rework→fail slope.

Honest magnitude: forcing the coarse discrete levers into the band would have
needed an unphysical ~10× rim. Instead the slope is bought with a continuous
refining lever — front_purity's k^n is smooth in n, so zone_passes relaxes to
float and a fractional pass lands the band integer passes leap over — letting a
modest, clearly-real 3.5× rim suffice. Both the r² profile and the steepness are
flagged house numbers (ADR 0005 §5), not derived.

Seam-exact: gated off when variation is disabled (uniform Na ⇒ demo_device) and
irrelevant to a clean Na=0 feed; composes with the focus bowl through a different
spec channel (Q_ox→V_t vs CD→NILS), so they stack not double-count; litho-rework
does not rescue it (only purifying does). Lifts the purification module's own
flagged "within-wafer contamination non-uniformity is out" scope edge, for the
Na channel only.

This is the consequence-model mechanism, not a player stage — reachable only via
a test probe today; the purification stage (the continuous lever as a UX knob,
multi-step refining, the consequence preview, the staged journey) is next.

Tests: fab_game/tests/test_na_ring.py (9) — the radial factor, the partial ring
sits outboard and traces to Q_ox, the all-pass seam without the edge trend, a
clean feed untouched, the ring grows inward monotonically, the fractional-pass
lever, defocus composition, rework-no-rescue. Full fab_game + chip suites green.

+++ fab_game/recipe.py
@@ -44,7 +44,10 @@ class PurificationKnobs:
     """
 
     grade: str = "clean"               # FEEDSTOCK_GRADES key — "clean" idealized baseline (the seam)
-    zone_passes: int = 1               # zone-refining passes (more = cleaner feed; the costly rework knob)
+    zone_passes: float = 1             # zone-refining passes/effort (more = cleaner feed; the costly rework
+    #                                    knob). CONTINUOUS — a fractional pass is a partial refining effort
+    #                                    (front_purity's k^n is smooth in n), so the residual Na can be dialed
+    #                                    into the marginal band instead of leaping ~1/k per integer pass.
 
 
 @dataclass(frozen=True)
